backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==1.10.13
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1